backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
gunicorn==21.2.0
python-decouple==3.8
django-filter==23.5
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
isort==5.13.2
coverage==7.3.4

